+++ gencode/docs/config.html
@@ -1387,6 +1387,92 @@
             
 
             
+<div class="accordion" id="accordionpointset_timestamp">
+    <div class="card">
+        <div class="card-header" id="headingpointset_timestamp">
+            <h2 class="mb-0">
+                <button class="btn btn-link property-name-button" type="button" data-toggle="collapse" data-target="#pointset_timestamp"
+                        aria-expanded="" aria-controls="pointset_timestamp" onclick="setAnchor('#pointset_timestamp')"><span class="property-name">timestamp</span></button>
+            </h2>
+        </div>
+
+        <div id="pointset_timestamp"
+             class="collapse property-definition-div" aria-labelledby="headingpointset_timestamp"
+             data-parent="#accordionpointset_timestamp">
+            <div class="card-body pl-5">
+
+    <div class="breadcrumbs">root
+        <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-right-short" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            <path
+                fill-rule="evenodd"
+                d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
+            />
+        </svg>
+    <a href="#pointset" onclick="anchorLink('pointset')">pointset</a>
+        <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-right-short" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            <path
+                fill-rule="evenodd"
+                d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
+            />
+        </svg>
+    <a href="#pointset_timestamp" onclick="anchorLink('pointset_timestamp')">timestamp</a></div><span class="badge badge-dark value-type">Type: string</span><br/>
+<span class="description"><p>RFC 3339 timestamp the configuration was generated</p>
+</span>
+            
+
+            
+            
+
+            <br/>
+<div class="badge badge-secondary">Example:</div>
+<br/><div id="pointset_timestamp_ex1" class="jumbotron examples"><div class="highlight"><pre><span></span><span class="s2">&quot;2019-01-17T14:02:29.364Z&quot;</span>
+</pre></div>
+</div>
+            </div>
+        </div>
+    </div>
+</div>
+<div class="accordion" id="accordionpointset_version">
+    <div class="card">
+        <div class="card-header" id="headingpointset_version">
+            <h2 class="mb-0">
+                <button class="btn btn-link property-name-button" type="button" data-toggle="collapse" data-target="#pointset_version"
+                        aria-expanded="" aria-controls="pointset_version" onclick="setAnchor('#pointset_version')"><span class="property-name">version</span></button>
+            </h2>
+        </div>
+
+        <div id="pointset_version"
+             class="collapse property-definition-div" aria-labelledby="headingpointset_version"
+             data-parent="#accordionpointset_version">
+            <div class="card-body pl-5">
+
+    <div class="breadcrumbs">root
+        <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-right-short" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            <path
+                fill-rule="evenodd"
+                d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
+            />
+        </svg>
+    <a href="#pointset" onclick="anchorLink('pointset')">pointset</a>
+        <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-arrow-right-short" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            <path
+                fill-rule="evenodd"
+                d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
+            />
+        </svg>
+    <a href="#pointset_version" onclick="anchorLink('pointset_version')">version</a></div><span class="badge badge-dark value-type">Type: integer</span><br/>
+<span class="description"><p>Major version of the UDMI schema</p>
+</span>
+            
+
+            
+            <p><span class="badge badge-light restriction numeric-restriction" id="pointset_version_number">Value must be greater or equal to <code>1</code> and lesser or equal to <code>1</code></span></p>
+
+            
+            </div>
+        </div>
+    </div>
+</div>
 <div class="accordion" id="accordionpointset_state_etag">
     <div class="card">
         <div class="card-header" id="headingpointset_state_etag">
