+++ gencode/docs/metadata.html
@@ -1452,7 +1452,7 @@
                 d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
             />
         </svg>
-    <a href="#localnet_subsystem_pattern1" onclick="anchorLink('localnet_subsystem_pattern1')">metadata_localnet_subsystem.json</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
+    <a href="#localnet_subsystem_pattern1" onclick="anchorLink('localnet_subsystem_pattern1')">model_localnet_subsystem.json</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
 <span class="description"><p>The type of network</p>
 </span>
 
@@ -1648,7 +1648,7 @@
                 d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
             />
         </svg>
-    <a href="#testing_targets_pattern1" onclick="anchorLink('testing_targets_pattern1')">metadata_testing_target.json#</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
+    <a href="#testing_targets_pattern1" onclick="anchorLink('testing_targets_pattern1')">model_testing_target.json#</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
 
 
      <span class="badge badge-info no-additional">No Additional Properties</span>
@@ -1889,7 +1889,7 @@
                 d="M4 8a.5.5 0 0 1 .5-.5h5.793L8.146 5.354a.5.5 0 1 1 .708-.708l3 3a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708-.708L10.293 8.5H4.5A.5.5 0 0 1 4 8z"
             />
         </svg>
-    <a href="#pointset_points_pattern1" onclick="anchorLink('pointset_points_pattern1')">metadata_pointset_point.json#</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
+    <a href="#pointset_points_pattern1" onclick="anchorLink('pointset_points_pattern1')">model_pointset_point.json#</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
 <span class="description"><p>Information about a specific point name of the device.</p>
 </span>
