+++ gencode/python/udmi/schema/__init__.py
@@ -15,19 +15,20 @@ from .event_discovery import DiscoveryEvent
 from .event_discovery_blob import BlobEnumerationEvent
 from .event_discovery_family import FamilyDiscoveryEvent
 from .event_discovery_point import PointEnumerationEvent
+from .event_mapping import DiscoveryEvent
 from .event_pointset import PointsetEvent
 from .event_pointset_point import PointPointsetEvent
 from .event_system import SystemEvent
 from .metadata import Metadata
-from .metadata_cloud import CloudMetadata
-from .metadata_gateway import GatewayMetadata
-from .metadata_localnet import LocalnetMetadata
-from .metadata_localnet_subsystem import SubsystemLocalnetMetadata
-from .metadata_pointset import PointsetMetadata
-from .metadata_pointset_point import PointPointsetMetadata
-from .metadata_system import SystemMetadata
-from .metadata_testing import TestingMetadata
-from .metadata_testing_target import TargetTestingMetadata
+from .model_cloud import CloudMetadata
+from .model_gateway import GatewayMetadata
+from .model_localnet import LocalnetMetadata
+from .model_localnet_subsystem import SubsystemLocalnetMetadata
+from .model_pointset import PointsetMetadata
+from .model_pointset_point import PointPointsetMetadata
+from .model_system import SystemMetadata
+from .model_testing import TestingMetadata
+from .model_testing_target import TargetTestingMetadata
 from .properties import Properties
 from .state import State
 from .state_blobset import BlobsetState
