+++ gencode/python/udmi/schema/metadata.py
@@ -1,10 +1,10 @@
 """Generated class for metadata.json"""
-from .metadata_cloud import CloudMetadata
-from .metadata_system import SystemMetadata
-from .metadata_gateway import GatewayMetadata
-from .metadata_localnet import LocalnetMetadata
-from .metadata_testing import TestingMetadata
-from .metadata_pointset import PointsetMetadata
+from .model_cloud import CloudMetadata
+from .model_system import SystemMetadata
+from .model_gateway import GatewayMetadata
+from .model_localnet import LocalnetMetadata
+from .model_testing import TestingMetadata
+from .model_pointset import PointsetMetadata
 
 
 class Metadata:
