+++ gencode/python/udmi/schema/model_localnet.py
@@ -1,22 +1,19 @@
-"""Generated class for metadata_gateway.json"""
+"""Generated class for model_localnet.json"""
+from .model_localnet_subsystem import SubsystemLocalnetMetadata
 
 
-class GatewayMetadata:
+class LocalnetMetadata:
   """Generated schema class"""
 
   def __init__(self):
-    self.gateway_id = None
     self.subsystem = None
-    self.proxy_ids = None
 
   @staticmethod
   def from_dict(source):
     if not source:
       return None
-    result = GatewayMetadata()
-    result.gateway_id = source.get('gateway_id')
-    result.subsystem = source.get('subsystem')
-    result.proxy_ids = source.get('proxy_ids')
+    result = LocalnetMetadata()
+    result.subsystem = SubsystemLocalnetMetadata.map_from(source.get('subsystem'))
     return result
 
   @staticmethod
@@ -25,7 +22,7 @@ class GatewayMetadata:
       return None
     result = {}
     for key in source:
-      result[key] = GatewayMetadata.from_dict(source[key])
+      result[key] = LocalnetMetadata.from_dict(source[key])
     return result
 
   @staticmethod
@@ -37,10 +34,6 @@ class GatewayMetadata:
 
   def to_dict(self):
     result = {}
-    if self.gateway_id:
-      result['gateway_id'] = self.gateway_id # 5
     if self.subsystem:
-      result['subsystem'] = self.subsystem # 5
-    if self.proxy_ids:
-      result['proxy_ids'] = self.proxy_ids # 1
+      result['subsystem'] = SubsystemLocalnetMetadata.expand_dict(self.subsystem) # 2
     return result
